Move type files to the type directory.
